+++ src/main/java/com/xiyiaoo/service/impl/ResourceServiceImpl.java
@@ -3,14 +3,19 @@
  */
 package com.xiyiaoo.service.impl;
 
+import com.xiyiaoo.constants.Constants;
 import com.xiyiaoo.dao.ResourceDao;
 import com.xiyiaoo.entity.Resource;
 import com.xiyiaoo.service.ResourceService;
 import com.xiyiaoo.util.CollectionUtil;
 import com.xiyiaoo.util.StringUtil;
 import com.xiyiaoo.util.WebUtil;
+import com.xiyiaoo.validation.group.Create;
+import com.xiyiaoo.validation.group.Query;
+import com.xiyiaoo.validation.group.Update;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 import java.util.Collections;
 import java.util.List;
@@ -38,24 +43,35 @@ public class ResourceServiceImpl implements ResourceService {
     }
 
     @Override
-    public void add(Resource resource) {
-        resource.setId(StringUtil.uuid());
+    @Validated({Create.class})
+    public int add(Resource resource) {
+        //resource.setId(StringUtil.uuid());
         resource.setCreator(WebUtil.getCurrentUserId());
-        resourceDao.add(resource);
+        return resourceDao.add(resource);
     }
 
     @Override
-    public void delete(Resource resource) {
-        resourceDao.delete(resource);
+    @Validated({Query.class})
+    public int delete(Resource resource) {
+        if (Constants.ROOT_ID.equals(resource.getId())){
+            return 0;
+        }
+        //TODO 级联删除
+        return resourceDao.delete(resource);
     }
 
     @Override
-    public void update(Resource resource) {
+    @Validated({Update.class})
+    public int update(Resource resource) {
+        if (Constants.ROOT_ID.equals(resource.getId())){
+            return 0;
+        }
         resource.setModifier(WebUtil.getCurrentUserId());
-        resourceDao.update(resource);
+        return resourceDao.update(resource);
     }
 
     @Override
+    @Validated({Query.class})
     public Resource get(Resource resource) {
         return resourceDao.get(resource);
     }
